add ua aware css generation for chome mobile
this adds css generation support for chrome mobile. previous version did
not provide any format. i guess that chrome for android supports
truetype and svg as well, but as local and woff are supported and
preferred this change should be fine and for what it’s worth: the google
font webservice provides the same format options.